ViewThemeUtils: move things that do not belong in this class
Color calculations go to ColorUtil, checking if dark theme goes to PlatformThemeUtil
